Added in the BuzzSaw ship
BuzzSaw: uses the new player base script for setting material, setting the base root and using the new thruster.
Ability is functional but needs some tweaking.
The ability cost is doubled when in the circle.
Max fuel amount is higher than other ships.
Added in the Use ability action mapping.